Add Authorization on additional Controller

+++ app/Http/Controllers/HobbyTagController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use App\Hobby;
 use App\Tag;
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\Gate;
 
 class HobbyTagController extends Controller
 {
@@ -22,6 +22,9 @@ class HobbyTagController extends Controller
 
     public function attachTag($hobby_id, $tag_id){
         $hobby = Hobby::find($hobby_id);
+        if(Gate::denies('connect_hobbyTag',$hobby)){
+            abort(403,'This is not your hobby.');
+        }
         $tag = Tag::find($tag_id);
         $hobby->tags()->attach($tag_id);
         return back()->with([
@@ -31,6 +34,9 @@ class HobbyTagController extends Controller
 
     public function detachTag($hobby_id, $tag_id){
         $hobby = Hobby::find($hobby_id);
+        if(Gate::denies('connect_hobbyTag',$hobby)){
+            abort(403,'This is not your hobby.');
+        }
         $tag = Tag::find($tag_id);
         $hobby->tags()->detach($tag_id);
         return back()->with([
